requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
paho-mqtt==1.5.0
pprint==0.1
PyYAML==5.2
Werkzeug==0.16.0
